+++ tests/integration/scenarios/idle.js
@@ -259,45 +259,45 @@ describe("initial idle state", () => {
     });
 
     it("should have a -1 manual audio bitrate by default", () => {
-      expect(player.getManualAudioBitrate()).to.equal(-1);
+      expect(player.getForcedAudioBitrate()).to.equal(-1);
     });
 
     it("should have a -1 manual video bitrate by default", () => {
-      expect(player.getManualVideoBitrate()).to.equal(-1);
+      expect(player.getForcedVideoBitrate()).to.equal(-1);
     });
 
     it("should update manual video bitrate when calling setVideoBitrate", () => {
-      const oldManual = player.getManualVideoBitrate();
+      const oldForced = player.getForcedVideoBitrate();
 
       player.setVideoBitrate(84);
-      expect(player.getManualVideoBitrate()).to.equal(84);
+      expect(player.getForcedVideoBitrate()).to.equal(84);
 
       player.setVideoBitrate(-1);
-      expect(player.getManualVideoBitrate()).to.equal(-1);
+      expect(player.getForcedVideoBitrate()).to.equal(-1);
 
       player.setVideoBitrate(0);
-      expect(player.getManualVideoBitrate()).to.equal(0);
+      expect(player.getForcedVideoBitrate()).to.equal(0);
 
-      player.setVideoBitrate(oldManual);
-      expect(player.getManualVideoBitrate()).to.equal(oldManual);
+      player.setVideoBitrate(oldForced);
+      expect(player.getForcedVideoBitrate()).to.equal(oldForced);
     });
 
     it("should update manual audio bitrate when calling setAudioBitrate", () => {
-      const oldManual = player.getManualAudioBitrate();
+      const oldForced = player.getForcedAudioBitrate();
 
       player.setAudioBitrate(84);
-      expect(player.getManualAudioBitrate()).to.equal(84);
+      expect(player.getForcedAudioBitrate()).to.equal(84);
       player.setAudioBitrate(-1);
-      expect(player.getManualAudioBitrate()).to.equal(-1);
+      expect(player.getForcedAudioBitrate()).to.equal(-1);
       player.setAudioBitrate(0);
-      expect(player.getManualAudioBitrate()).to.equal(0);
+      expect(player.getForcedAudioBitrate()).to.equal(0);
 
-      player.setAudioBitrate(oldManual);
-      expect(player.getManualAudioBitrate()).to.equal(oldManual);
+      player.setAudioBitrate(oldForced);
+      expect(player.getForcedAudioBitrate()).to.equal(oldForced);
     });
 
     it("should update the maximum video bitrate when calling setMaxVideoBitrate by default", () => {
-      const oldMax = player.getManualVideoBitrate();
+      const oldMax = player.getForcedVideoBitrate();
 
       expect(player.setMaxVideoBitrate(Infinity)).to.equal(undefined);
       expect(player.getMaxVideoBitrate()).to.equal(Infinity);
@@ -316,7 +316,7 @@ describe("initial idle state", () => {
     });
 
     it("should update the maximum audio bitrate when calling setMaxAudioBitrate by default", () => {
-      const oldMax = player.getManualAudioBitrate();
+      const oldMax = player.getForcedAudioBitrate();
 
       expect(player.setMaxAudioBitrate(Infinity)).to.equal(undefined);
       expect(player.getMaxAudioBitrate()).to.equal(Infinity);